starting double -> float conversion

+++ src/architecture/msgPayloadDefC/ArrayMotorTorqueMsgPayload.h
@@ -27,7 +27,7 @@
 
 /*! @brief Structure used to define the output definition for vehicle effectors*/
 typedef struct {
-    double motorTorque[MAX_EFF_CNT];     //!< [Nm]  motor torque array
+    float motorTorque[MAX_EFF_CNT];     //!< [Nm]  motor torque array
 }ArrayMotorTorqueMsgPayload;
