+++ src/architecture/msgPayloadDefC/AttRefMsgPayload.h
@@ -24,9 +24,9 @@
 
 /*! @brief Structure used to define the output definition for attitude reference generation */
 typedef struct {
-    double sigma_RN[3];         //!<        MRP Reference attitude of R relative to N
-    double omega_RN_N[3];       //!< [r/s]  Reference frame rate vector of the of R relative to N in N frame components
-    double domega_RN_N[3];      //!< [r/s2] Reference frame inertial acceleration of  R relative to N in N frame components
+    float sigma_RN[3];         //!<        MRP Reference attitude of R relative to N
+    float omega_RN_N[3];       //!< [r/s]  Reference frame rate vector of the of R relative to N in N frame components
+    float domega_RN_N[3];      //!< [r/s2] Reference frame inertial acceleration of  R relative to N in N frame components
 }AttRefMsgPayload;
