+++ src/architecture/msgPayloadDefC/CSSArraySensorMsgPayload.h
@@ -25,7 +25,7 @@
 
 /*! @brief Output structure for CSS array or constellation interface.  Each element contains the raw measurement which should be a cosine value nominally */
 typedef struct {
-    double CosValue[MAX_NUM_CSS_SENSORS];   //!< Current measured CSS value (ideally a cosine value) for the constellation of CSS sensors
+    float CosValue[MAX_NUM_CSS_SENSORS];   //!< Current measured CSS value (ideally a cosine value) for the constellation of CSS sensors
 }CSSArraySensorMsgPayload;
 
 #endif
