+++ src/architecture/msgPayloadDefCpp/OpNavUnitVecMsgPayload.h
@@ -29,14 +29,14 @@ typedef struct
 OpNavUnitVecMsgPayload
 //@endcond
 {
-    double timeTag;                             //!< [s] Current time of validity for output
+    float timeTag;                             //!< [s] Current time of validity for output
     bool valid;     //!< Quality of measurement if 1, invalid if 0
-    double covar_N[3*3];    //!< [-] Covariance of measurement in the inertial frame
-    double covar_B[3*3];    //!< [-] Covariance of measurement in the body frame
-    double covar_C[3*3];    //!< [-] Covariance of measurement in the camera frame
-    double rhat_BN_N[3];                 //!< [-] measurement in the inertial frame
-    double rhat_BN_B[3];                 //!< [-] measurement in the body frame
-    double rhat_BN_C[3];                 //!< [-] measurement in the camera frame
+    float covar_N[3*3];    //!< [-] Covariance of measurement in the inertial frame
+    float covar_B[3*3];    //!< [-] Covariance of measurement in the body frame
+    float covar_C[3*3];    //!< [-] Covariance of measurement in the camera frame
+    float rhat_BN_N[3];                 //!< [-] measurement in the inertial frame
+    float rhat_BN_B[3];                 //!< [-] measurement in the body frame
+    float rhat_BN_C[3];                 //!< [-] measurement in the camera frame
 }OpNavUnitVecMsgPayload;
 
 #endif /* OPNAVUVECMESSAGE_H */
